+++ source/src/nemo-action-config-widget.c
@@ -366,8 +366,7 @@ on_open_folder_clicked (GtkWidget *button, NemoActionConfigWidget *widget)
     nemo_application_open_location (nemo_application_get_singleton (),
                                     location,
                                     NULL,
-                                    NEMO_APP_SLUG,
-                                    FALSE);
+                                    NEMO_APP_SLUG);
 
     g_free (path);
     g_object_unref (location);