+++ source/src/nemo-location-bar.c
@@ -146,7 +146,7 @@ drag_data_received_callback (GtkWidget *widget,
 	char **names;
 	NemoApplication *application;
 	int name_count;
-	NemoWindow *new_window, *window;
+	NemoWindow *window;
 	GdkScreen      *screen;
 	gboolean new_windows_for_extras;
 	char *prompt;
@@ -213,9 +213,8 @@ drag_data_received_callback (GtkWidget *widget,
 		screen = gtk_window_get_screen (GTK_WINDOW (window));
 
 		for (i = 1; names[i] != NULL; ++i) {
-			new_window = nemo_application_create_window (application, screen);
 			location = g_file_new_for_uri (names[i]);
-			nemo_window_go_to (new_window, location);
+			nemo_application_open_in_new_window (application, screen, location, NULL);
 			g_object_unref (location);
 		}
 	}